Extracting common functions to be shared
